requirements.txt
alembic==1.7.7
appnope==0.1.3
asttokens==2.0.5
attrs==21.4.0
autopage==0.5.0
autopep8==1.6.0
backcall==0.2.0
bitstring==3.1.9
certifi==2021.10.8
charset-normalizer==2.0.12
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorlog==6.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
filprofiler==2022.3.0
flake8==4.0.1
fonttools==4.32.0
greenlet==1.1.2
h5py==3.6.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jaconv==0.3
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.5
numpy==1.22.2
optuna==2.10.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prettytable==3.2.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyDeprecate==0.3.2
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.8
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
stevedore==3.5.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.11.0
torchmetrics==0.8.0
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
